+++ src/Modules/Admin/Widgets/Forms/AccountCredentialsActiveForm.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Hirtz\Skeleton\Modules\Admin\Widgets\Forms;
 
-use Hirtz\Skeleton\Models\Forms\AccountUpdateForm;
+use Hirtz\Skeleton\Models\Forms\AccountCredentialsForm;
 use Hirtz\Skeleton\Modules\Admin\Controllers\AccountController;
 use Hirtz\Skeleton\Modules\Admin\Widgets\Forms\Traits\UserActiveFormTrait;
 use Hirtz\Skeleton\Widgets\Forms\ActiveForm;
@@ -15,7 +15,7 @@ use Stringable;
 /**
  * @see AccountController::actionCredentials()
  *
- * @property AccountUpdateForm $model
+ * @property AccountCredentialsForm $model
  */
 class AccountCredentialsActiveForm extends ActiveForm
 {